SystemUI: QS: Add pulldown with one finger (RTL aware) (Squashed)
* A squashed commit of all the commits below
* RTL aware, setting value exposed in config.xml (autonomous, no Settings commit required)
* Adapted for Marshmallow by @Kreach3R

Logic is currently set to use 1/3 of right hand side of the statusbar to
trigger a quick settings quick pull down, quickly.

Change-Id: I3f536496f724f07d2419a5b2db5fa8f8261609e7
Signed-off-by: Roman Birg <roman@cyngn.com>

SystemUI: fix right hand side gestures on lockscreen

two finger quick pulldown was not accounting for being in the keyguard
adn was eating touch events when it shouldn't be.

Change-Id: I625d48a4002aaaf2d55d7efaed65d892ab7aa13c
Signed-off-by: Roman Birg <roman@cyngn.com>

SystemUI: Reduce activation area a bit of quick pulldown

Change-Id: I06b3a50982930491c6fd4e70ceda3b1ceee28f87

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java

FB: add ability to use quick settings pulldown on the left side (value set in config.xml)

This commit gives the ability to use quick settings pulldown on the left
or right edge of the statusbar. As well this respects still RTL layouts
where it logically should change to the oposite side.

Change-Id: Ibdf0894347e1ed3871f685afd5a29f47efe23a4b

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java

fixup pulldown